Reorganize file structure for better clarity and separation of concerns. Moved core image processing logic to `src/core/` and renamed the application-specific directory to `src/app/`. Updated import statements to reflect these changes.
