+++ src/app/image_reader.py
