+++ src/core/image_processor.py
